[MIG]osi_analytic_segments_defaults migrated in v14.
