+++ osi_analytic_segments_defaults/i18n/en_US.po
@@ -30,7 +30,7 @@ msgstr "Analytic Segment One"
 msgid "Analytic Segment Two"
 msgstr "Analytic Segment Two"
 
-#. module: account_analytic_default
+#. module: account
 #: model:ir.model.fields,field_description:account_analytic_default.field_account_analytic_default_analytic_id
 #: model:ir.ui.view,arch_db:account_analytic_default.view_account_analytic_default_form_search
 msgid "Analytic Account"
