+++ osi_analytic_segments_defaults/models/__init__.py
@@ -1,7 +1,7 @@
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 from . import (
     account_analytic_default,
-    account_invoice_line,
+    account_move_line,
     purchase_order,
     sale_order_line,
 )
